Round one of class cleanup
