+++ src/Command/DockerCommands.php
@@ -2,6 +2,9 @@
 
 namespace DkanTools\Command;
 
+use DkanTools\Util\Util;
+use Robo\Task;
+
 /**
  * These functions are not meant to be run, simply to provide documentation for
  * the docker comands supplied by dktl.sh.
@@ -27,6 +30,8 @@ class DockerCommands extends \Robo\Tasks
      *
      * You may also run a shorter alias, "dc" (e.g. "dktl dc ps"). See
      * https://docs.docker.com/compose/reference/ for a full list of commands.
+     *
+     * @aliases dc
      */
     public function dockerCompose()
     {
@@ -34,33 +39,36 @@ class DockerCommands extends \Robo\Tasks
     }
 
     /**
-     * Display the insecure (http) web URL for the current project.
+     * Display the http web URL for the current project.
+     *
+     * @aliases url
      */
-    public function url()
+    public function dockerUrl()
+    {
+        $uri = Util::getUri();
+        if ($uri) {
+            return $uri;
+        } else {
+            throw new \Exception("URI not available!");
+        }
+    }
+
+    /**
+     * Connect the web container to the proxy.
+     *
+     * @aliases proxy:connect
+     */
+    public function dockerProxyConnect()
     {
         throw new \Exception(self::DKTL_DOCKER_PHP_ERROR);
     }
 
     /**
-     * Display the secure (https) web URL for the current project.
+     * Kill the dktl proxy service.
+     *
+     * @aliases proxy:kill
      */
-    public function surl()
-    {
-        throw new \Exception(self::DKTL_DOCKER_PHP_ERROR);
-    }
-
-    /**
-     * Show clickable admin login link for insecure web URL (http)
-     */
-    public function uli()
-    {
-        throw new \Exception(self::DKTL_DOCKER_PHP_ERROR);
-    }
-
-    /**
-     * Show clickable admin login link for secure URL (https)
-     */
-    public function suli()
+    public function dockerProxyKill()
     {
         throw new \Exception(self::DKTL_DOCKER_PHP_ERROR);
     }
