+++ src/SymlinksTrait.php
@@ -7,6 +7,8 @@ use Symfony\Component\Filesystem\Filesystem;
 
 trait SymlinksTrait
 {
+    use DrupalProjectTrait;
+
     private function addSymlinksToDrupalRoot()
     {
         $targetsAndLinks = [
@@ -18,7 +20,7 @@ trait SymlinksTrait
         foreach ($targetsAndLinks as $targetAndLink) {
             $this->docrootSymlink(
                 $targetAndLink['target'],
-                self::DRUPAL_FOLDER_NAME . $targetAndLink['link']
+                self::$drupalDocroot . $targetAndLink['link']
             );
         }
     }
@@ -32,10 +34,10 @@ trait SymlinksTrait
         $link_dirname = $link_parts['dirname'];
         $target_path_relative_to_link = (new Filesystem())->makePathRelative($target, $link_dirname);
 
-        if (!file_exists($target) || !file_exists(self::DRUPAL_FOLDER_NAME)) {
+        if (!file_exists($target) || !file_exists(self::$drupalDocroot)) {
             $this->io()->warning(
                 "Skipping linking $target. Folders $target and '" .
-                self::DRUPAL_FOLDER_NAME."' must both be present to create link."
+                self::$drupalDocroot . "' must both be present to create link."
             );
             return;
         }
